server/requirements.txt
aiohttp==3.5.4
alembic==1.0.10
appnope @ file:///opt/concourse/worker/volumes/live/0291c9e1-4b15-459f-623e-2770f55be269/volume/appnope_1594338395037/work
APScheduler==3.6.1
argon2-cffi @ file:///opt/concourse/worker/volumes/live/59af29ac-4890-416e-7ab7-794f8d6f7ecd/volume/argon2-cffi_1596828548321/work
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
backcall==0.2.0
backports-datetime-fromisoformat==1.0.0
bcrypt==3.1.6
beautifulsoup4==4.9.3
bleach==3.1.5
blinker==1.4
blueprint==3.4.2
boto3==1.9.134
botocore==1.12.134
brotlipy==0.7.0
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
conda==4.8.5
conda-package-handling==1.7.0+0.g7c4a471.dirty
coolsms-python-sdk==2.0.3
country-converter==0.6.7
coverage==5.1
cryptography @ file:///opt/concourse/worker/volumes/live/a7ee5cc5-41f4-40e1-6671-3e164687c6e0/volume/cryptography_1598892026615/work
customerio==0.5.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
firebase-admin==4.5.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Env==2.0.0
Flask-JWT-Extended==3.18.1
flask-marshmallow==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-language==2.0.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==0.4.15
grpcio==1.36.1
gunicorn==19.9.0
httplib2==0.19.0
idna==2.8
importlib-metadata==1.6.0
ipykernel @ file:///opt/concourse/worker/volumes/live/88f541d3-5a27-498f-7391-f2e50ca36560/volume/ipykernel_1596206680118/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///opt/concourse/worker/volumes/live/911d827b-e6b1-486b-686d-e48c633263da/volume/ipython_1598371556001/work
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi @ file:///opt/concourse/worker/volumes/live/c63c70ef-654a-4d26-6b78-37aef096f225/volume/jedi_1596490811751/work
Jinja2==2.10
jmespath==0.9.4
jsonschema @ file:///opt/concourse/worker/volumes/live/143dedb7-a810-4064-6e47-436d4dbcc943/volume/jsonschema_1594363611216/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.1
libcst==0.3.17
lxml==4.6.2
Mako==1.0.9
MarkupSafe @ file:///opt/concourse/worker/volumes/live/cb778296-98db-45ad-411e-6f726e102dc3/volume/markupsafe_1594371638608/work
marshmallow==2.19.1
marshmallow-sqlalchemy==0.16.1
matplotlib==3.2.1
mccabe==0.6.1
MechanicalSoup==1.0.0
mistune @ file:///opt/concourse/worker/volumes/live/95802d64-d39c-491b-74ce-b9326880ca54/volume/mistune_1594373201816/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.2.0
msgpack==1.0.2
multidict==4.5.2
mypy-extensions==0.4.3
nbconvert @ file:///opt/concourse/worker/volumes/live/70406b39-619d-477e-7507-0cc36a02f91a/volume/nbconvert_1594376816864/work
nbformat==5.0.7
notebook @ file:///opt/concourse/worker/volumes/live/3ff7115c-cc67-4f6b-6e2e-c0806c78c21d/volume/notebook_1596838654712/work
numpy==1.18.2
olefile==0.46
opencv-python==4.2.0.34
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect @ file:///opt/concourse/worker/volumes/live/8701bb20-ad87-46c7-5108-30c178cf97e5/volume/pexpect_1594383388344/work
pickleshare @ file:///opt/concourse/worker/volumes/live/93ec39d8-05bb-4f84-7efc-98735bc39b70/volume/pickleshare_1594384101884/work
Pillow @ file:///opt/concourse/worker/volumes/live/be1e8a56-c4be-4ffe-4fa7-5a0e9c460b1a/volume/pillow_1594307312933/work
pluggy==0.13.1
postgres==2.2.2
pprintpp==0.4.0
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598885458782/work
proto-plus==1.14.2
protobuf==3.15.5
psycopg2==2.8.2
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.1
pyactiveresource==2.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycountry==19.8.18
pycountry-convert==0.7.2
pycparser==2.19
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.11.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks @ file:///opt/concourse/worker/volumes/live/85a5b906-0e08-41d9-6f59-084cee4e9492/volume/pysocks_1594394636991/work
pytest==5.4.1
pytest-cov==2.8.1
pytest-mock==3.0.0
python-dateutil==2.8.0
python-editor==1.0.4
python-http-client==3.1.0
pytz==2019.1
PyYAML==5.4.1
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1598374667791/work
QtPy==1.9.0
redis==3.5.3
repoze.lru==0.7
requests==2.22.0
rsa==4.7.2
ruamel-yaml==0.15.87
s3transfer==0.2.0
scipy @ file:///opt/concourse/worker/volumes/live/7f5e51f8-abf3-4409-4c56-a4fb14cc9869/volume/scipy_1593104185999/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
selenium==3.141.0
Send2Trash==1.5.0
sendgrid==6.0.5
sentry-sdk==0.14.2
ShopifyAPI==5.1.2
six==1.15.0
slackclient==2.1.0
soupsieve==1.9.1
SQLAlchemy==1.3.1
sqlparse==0.3.0
stripe==2.29.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.24.2
virtualenv==16.7.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.3.0
zipp==3.1.0
